refactor ticket models (#1106)
